+++ ios/NimbleEdgeAI/Utils/AsyncSemaphore.swift
@@ -1,0 +1,27 @@
+actor AsyncSemaphore {
+    private var value: Int
+    private var waitQueue: [CheckedContinuation<Void, Never>] = []
+
+    init(value: Int) {
+        self.value = value
+    }
+
+    func wait() async {
+        if value > 0 {
+            value -= 1
+        } else {
+            await withCheckedContinuation { continuation in
+                waitQueue.append(continuation)
+            }
+        }
+    }
+
+    func signal() {
+        if !waitQueue.isEmpty {
+            let continuation = waitQueue.removeFirst()
+            continuation.resume()
+        } else {
+            value += 1
+        }
+    }
+}
